- split save mail & server out to separate files & other refactoring
