+++ util.go
@@ -1,0 +1,172 @@
+package main
+
+import (
+	"strings"
+	"crypto/md5"
+	"encoding/hex"
+	"errors"
+	"regexp"
+	"bytes"
+	"encoding/base64"
+	"io/ioutil"
+	"github.com/sloonz/go-qprintable"
+	"compress/zlib"
+	"gopkg.in/iconv.v1"
+)
+
+func validateEmailData(client *Client) (user string, host string, addr_err error) {
+	if user, host, addr_err = extractEmail(client.mail_from); addr_err != nil {
+		return user, host, addr_err
+	}
+	client.mail_from = user + "@" + host
+	if user, host, addr_err = extractEmail(client.rcpt_to); addr_err != nil {
+		return user, host, addr_err
+	}
+	client.rcpt_to = user + "@" + host
+	// check if on allowed hosts
+	if allowed := allowedHosts[strings.ToLower(host)]; !allowed {
+		return user, host, errors.New("invalid host:" + host)
+	}
+	return user, host, addr_err
+}
+
+func extractEmail(str string) (name string, host string, err error) {
+	re, _ := regexp.Compile(`<(.+?)@(.+?)>`) // go home regex, you're drunk!
+	if matched := re.FindStringSubmatch(str); len(matched) > 2 {
+		host = validHost(matched[2])
+		name = matched[1]
+	} else {
+		if res := strings.Split(str, "@"); len(res) > 1 {
+			name = res[0]
+			host = validHost(res[1])
+		}
+	}
+	if host == "" || name == "" {
+		err = errors.New("Invalid address, [" + name + "@" + host + "] address:" + str)
+	}
+	return name, host, err
+}
+
+// Decode strings in Mime header format
+// eg. =?ISO-2022-JP?B?GyRCIVo9dztSOWJAOCVBJWMbKEI=?=
+func mimeHeaderDecode(str string) string {
+	reg, _ := regexp.Compile(`=\?(.+?)\?([QBqp])\?(.+?)\?=`)
+	matched := reg.FindAllStringSubmatch(str, -1)
+	var charset, encoding, payload string
+	if matched != nil {
+		for i := 0; i < len(matched); i++ {
+			if len(matched[i]) > 2 {
+				charset = matched[i][1]
+				encoding = strings.ToUpper(matched[i][2])
+				payload = matched[i][3]
+				switch encoding {
+				case "B":
+					str = strings.Replace(
+						str,
+						matched[i][0],
+						mailTransportDecode(payload, "base64", charset),
+						1)
+				case "Q":
+					str = strings.Replace(
+						str,
+						matched[i][0],
+						mailTransportDecode(payload, "quoted-printable", charset),
+						1)
+				}
+			}
+		}
+	}
+	return str
+}
+
+func validHost(host string) string {
+	host = strings.Trim(host, " ")
+	re, _ := regexp.Compile(`^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9\-]*[A-Za-z0-9])$`)
+	if re.MatchString(host) {
+		return host
+	}
+	return ""
+}
+
+// decode from 7bit to 8bit UTF-8
+// encoding_type can be "base64" or "quoted-printable"
+func mailTransportDecode(str string, encoding_type string, charset string) string {
+	if charset == "" {
+		charset = "UTF-8"
+	} else {
+		charset = strings.ToUpper(charset)
+	}
+	if encoding_type == "base64" {
+		str = fromBase64(str)
+	} else if encoding_type == "quoted-printable" {
+		str = fromQuotedP(str)
+	}
+
+	if charset != "UTF-8" {
+		charset = fixCharset(charset)
+		if cd, err := iconv.Open("UTF-8", charset); err == nil {
+			defer func() {
+				cd.Close()
+				if r := recover(); r != nil {
+					//logln(1, fmt.Sprintf("Recovered in %v", r))
+				}
+			}()
+			// eg. charset can be "ISO-2022-JP"
+			return cd.ConvString(str)
+		}
+
+	}
+	return str
+}
+
+func fromBase64(data string) string {
+	buf := bytes.NewBufferString(data)
+	decoder := base64.NewDecoder(base64.StdEncoding, buf)
+	res, _ := ioutil.ReadAll(decoder)
+	return string(res)
+}
+
+func fromQuotedP(data string) string {
+	buf := bytes.NewBufferString(data)
+	decoder := qprintable.NewDecoder(qprintable.BinaryEncoding, buf)
+	res, _ := ioutil.ReadAll(decoder)
+	return string(res)
+}
+
+func compress(s string) string {
+	var b bytes.Buffer
+	w, _ := zlib.NewWriterLevel(&b, zlib.BestSpeed) // flate.BestCompression
+	w.Write([]byte(s))
+	w.Close()
+	return b.String()
+}
+
+func fixCharset(charset string) string {
+	reg, _ := regexp.Compile(`[_:.\/\\]`)
+	fixed_charset := reg.ReplaceAllString(charset, "-")
+	// Fix charset
+	// borrowed from http://squirrelmail.svn.sourceforge.net/viewvc/squirrelmail/trunk/squirrelmail/include/languages.php?revision=13765&view=markup
+	// OE ks_c_5601_1987 > cp949
+	fixed_charset = strings.Replace(fixed_charset, "ks-c-5601-1987", "cp949", -1)
+	// Moz x-euc-tw > euc-tw
+	fixed_charset = strings.Replace(fixed_charset, "x-euc", "euc", -1)
+	// Moz x-windows-949 > cp949
+	fixed_charset = strings.Replace(fixed_charset, "x-windows_", "cp", -1)
+	// windows-125x and cp125x charsets
+	fixed_charset = strings.Replace(fixed_charset, "windows-", "cp", -1)
+	// ibm > cp
+	fixed_charset = strings.Replace(fixed_charset, "ibm", "cp", -1)
+	// iso-8859-8-i -> iso-8859-8
+	fixed_charset = strings.Replace(fixed_charset, "iso-8859-8-i", "iso-8859-8", -1)
+	if charset != fixed_charset {
+		return fixed_charset
+	}
+	return charset
+}
+
+func md5hex(str string) string {
+	h := md5.New()
+	h.Write([]byte(str))
+	sum := h.Sum([]byte{})
+	return hex.EncodeToString(sum)
+}
